parameterise the owner and group
Pulled in from [charges-borrower-frontend#320a899](https://github.com/LandRegistry/charges-borrower-frontend/commit/320a899b76924e5846b8d2b95a427489676d9a6d).

When running the app in vagrant we would prefer the vagrant user to run
the app, so that when we ssh in we can easily interact with the machine.

When running in the environment we should enable the apps to be run as
their own dedicated user (as old school ops-ers recommend).

To enable this I have parameterised the owner and group and defaulted
them back to vagrant so there is no functional difference.